requirements.txt
Django
pandas
openpyxl
djangorestframework